+++ Assets/Scripts/Status.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class Status : MonoBehaviour
 {
-    public int VidaInicial = 100;
-    public int Vida;
+    public float VidaInicial = 100;
+    public float Vida;
     public float Velocidade = 5;
     public float attack = 1;
-    public float armor = 1;
+    public int armor = 1;
     public float attackSpeed = 0;
 
     void Awake ()
